feat: Track file chunking progress
Introduces a new `chunkingProgress` field to the state to track the progress of the worker chunking the input file. This allows for better user feedback during the analysis preparation phase.

Also includes minor improvements to file type handling in the FileUploadArea and resolves an issue with loading external libraries in the web worker.